# Simplify the bulk load algorithm