… rtree/golden_test.go