## Description

The previous bulk loading algorithm needed to keep the number of levels consistent in the r-tree consistent (i.e. every leaf node had to be the same distance from the root). This as a constraint required for the `Insert` and `Delete` methods. Because those methods no longer exist, we can greatly simplify the bulk loading algorithm.

With the simplified algorithm, the distance between the root and the leaf nodes will differ by at most 1 level. Nodes are now packed much more tightly, which reduces memory usage for some trees (but DOESN'T improve CPU performance).

## Check List

Have you:

- Added unit tests? N/A

- Add cmprefimpl tests? (if appropriate?) N/A

- Updated release notes? (if appropriate?) Yes.

## Related Issue

- N/A

## Benchmark Results

Benchmarks are basically as expected -- less memory usage and CPU about the same (some faster, some slower, but nothing significant).

<details>

<summary>Click to expand</summary>

```
goos: linux
goarch: arm64
pkg: github.com/peterstace/simplefeatures/geom
                                                           │ /tmp/tmp.9ynJa1Wrvi │         /tmp/tmp.fiLFr6Ksvk         │
                                                           │       sec/op        │    sec/op     vs base               │
LineEnvelope/0-2                                                   2.115n ±   4%   2.126n ±  1%       ~ (p=0.506 n=15)
LineEnvelope/1-2                                                   2.110n ±   1%   2.112n ±  1%       ~ (p=0.665 n=15)
LineEnvelope/2-2                                                   2.090n ±   0%   2.091n ±  0%       ~ (p=0.615 n=15)
LineEnvelope/3-2                                                   2.091n ±   1%   2.100n ±  1%  +0.43% (p=0.035 n=15)
MarshalWKB/polygon/n=10-2                                          92.45n ±   1%   92.38n ±  1%       ~ (p=0.616 n=15)
MarshalWKB/polygon/n=100-2                                         239.7n ±   3%   235.0n ±  2%  -1.96% (p=0.034 n=15)
MarshalWKB/polygon/n=1000-2                                        1.614µ ±   4%   1.634µ ± 15%       ~ (p=0.595 n=15)
MarshalWKB/polygon/n=10000-2                                       48.15µ ±  36%   43.26µ ± 19%       ~ (p=0.067 n=15)
UnmarshalWKB/polygon/n=10-2                                        158.1n ±   2%   157.5n ±  1%       ~ (p=0.545 n=15)
UnmarshalWKB/polygon/n=100-2                                       313.3n ±   4%   303.3n ±  2%       ~ (p=0.080 n=15)
UnmarshalWKB/polygon/n=1000-2                                      1.844µ ±   7%   1.818µ ±  5%       ~ (p=0.351 n=15)
UnmarshalWKB/polygon/n=10000-2                                     45.20µ ±  16%   45.93µ ± 15%       ~ (p=0.775 n=15)
IntersectsLineStringWithLineString/n=10-2                          734.1n ±   1%   711.1n ±  1%  -3.13% (p=0.000 n=15)
IntersectsLineStringWithLineString/n=100-2                         11.48µ ±   2%   10.43µ ±  1%  -9.17% (p=0.000 n=15)
IntersectsLineStringWithLineString/n=1000-2                        131.8µ ±   7%   137.2µ ±  4%       ~ (p=0.486 n=15)
IntersectsLineStringWithLineString/n=10000-2                       2.133m ±  10%   2.118m ± 14%       ~ (p=0.902 n=15)
IntersectsMultiPointWithMultiPoint/n=20-2                          555.8n ±   1%   555.3n ±  1%       ~ (p=0.546 n=15)
IntersectsMultiPointWithMultiPoint/n=200-2                         7.503µ ±   0%   7.490µ ±  0%       ~ (p=0.091 n=15)
IntersectsMultiPointWithMultiPoint/n=2000-2                        71.32µ ±   2%   71.40µ ±  1%       ~ (p=0.486 n=15)
IntersectsMultiPointWithMultiPoint/n=20000-2                       779.3µ ±   0%   777.0µ ±  0%       ~ (p=0.187 n=15)
PolygonSingleRingValidation/n=10-2                                 1.259µ ±   1%   1.236µ ±  1%  -1.83% (p=0.000 n=15)
PolygonSingleRingValidation/n=100-2                                18.80µ ±   1%   17.69µ ±  2%  -5.90% (p=0.000 n=15)
PolygonSingleRingValidation/n=1000-2                               272.3µ ±   0%   282.5µ ±  1%  +3.75% (p=0.000 n=15)
PolygonSingleRingValidation/n=10000-2                              3.383m ±   1%   3.411m ±  1%  +0.85% (p=0.002 n=15)
PolygonMultipleRingsValidation/n=4-2                               3.612µ ±   0%   3.468µ ±  0%  -3.99% (p=0.000 n=15)
PolygonMultipleRingsValidation/n=36-2                              30.55µ ±   1%   29.45µ ±  1%  -3.63% (p=0.000 n=15)
PolygonMultipleRingsValidation/n=400-2                             401.1µ ±   1%   384.3µ ±  1%  -4.19% (p=0.000 n=15)
PolygonMultipleRingsValidation/n=4096-2                            5.023m ±   1%   4.560m ±  1%  -9.22% (p=0.000 n=15)
PolygonZigZagRingsValidation/n=10-2                                5.996µ ±   0%   5.910µ ±  0%  -1.43% (p=0.002 n=15)
PolygonZigZagRingsValidation/n=100-2                               72.23µ ±   1%   69.29µ ±  1%  -4.08% (p=0.000 n=15)
PolygonZigZagRingsValidation/n=1000-2                              984.2µ ±   2%   981.6µ ±  3%       ~ (p=0.233 n=15)
PolygonZigZagRingsValidation/n=10000-2                             12.55m ±   2%   12.61m ±  2%       ~ (p=0.187 n=15)
PolygonAnnulusValidation/n=10-2                                    1.928µ ±   2%   1.863µ ±  0%  -3.37% (p=0.000 n=15)
PolygonAnnulusValidation/n=100-2                                   18.82µ ±   1%   18.73µ ±  1%       ~ (p=0.061 n=15)
PolygonAnnulusValidation/n=1000-2                                  329.0µ ±   1%   303.7µ ±  0%  -7.67% (p=0.000 n=15)
PolygonAnnulusValidation/n=10000-2                                 4.241m ±   2%   4.257m ±  1%       ~ (p=0.116 n=15)
MultipolygonValidation/n=1-2                                       212.6n ±   1%   199.3n ±  2%  -6.26% (p=0.000 n=15)
MultipolygonValidation/n=4-2                                       481.2n ±   1%   476.0n ±  1%  -1.08% (p=0.016 n=15)
MultipolygonValidation/n=16-2                                      2.092µ ±   0%   2.088µ ±  1%       ~ (p=0.454 n=15)
MultipolygonValidation/n=64-2                                      10.13µ ±   1%   10.19µ ±  1%       ~ (p=0.174 n=15)
MultipolygonValidation/n=256-2                                     56.31µ ±   2%   59.39µ ±  2%  +5.46% (p=0.000 n=15)
MultipolygonValidation/n=1024-2                                    338.1µ ±   1%   343.8µ ±  3%       ~ (p=0.595 n=15)
MultiPolygonTwoCircles/n=10-2                                      1.839µ ±   0%   1.817µ ±  1%  -1.20% (p=0.000 n=15)
MultiPolygonTwoCircles/n=100-2                                     22.08µ ±   1%   20.23µ ±  1%  -8.38% (p=0.000 n=15)
MultiPolygonTwoCircles/n=1000-2                                    263.4µ ±   5%   279.0µ ±  3%  +5.92% (p=0.001 n=15)
MultiPolygonTwoCircles/n=10000-2                                   4.614m ±  16%   5.107m ±  6%       ~ (p=0.412 n=15)
MultiPolygonMultipleTouchingPoints/n=1-2                           2.558µ ±   1%   2.528µ ±  1%  -1.17% (p=0.032 n=15)
MultiPolygonMultipleTouchingPoints/n=10-2                          20.65µ ±   2%   20.45µ ±  2%  -0.97% (p=0.045 n=15)
MultiPolygonMultipleTouchingPoints/n=100-2                         259.2µ ±   0%   258.4µ ±  1%  -0.32% (p=0.002 n=15)
MultiPolygonMultipleTouchingPoints/n=1000-2                        3.287m ±   2%   3.204m ±  2%  -2.54% (p=0.000 n=15)
WKTParsing/point-2                                                 1.133µ ±   2%   1.128µ ±  4%       ~ (p=0.631 n=15)
DistancePolygonToPolygonOrdering/n=100_swap=false-2                26.45µ ±   1%   24.86µ ±  2%  -6.01% (p=0.000 n=15)
DistancePolygonToPolygonOrdering/n=100_swap=true-2                 26.38µ ±   1%   24.87µ ±  2%  -5.73% (p=0.000 n=15)
DistancePolygonToPolygonOrdering/n=1000_swap=false-2               466.1µ ±   1%   462.5µ ±  0%       ~ (p=0.063 n=15)
DistancePolygonToPolygonOrdering/n=1000_swap=true-2                464.5µ ±   2%   461.9µ ±  0%  -0.57% (p=0.004 n=15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-2          3.195µ ±   1%   3.121µ ±  1%  -2.32% (p=0.000 n=15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-2           3.194µ ±   1%   3.111µ ±  1%  -2.60% (p=0.000 n=15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-2         34.34µ ±   1%   33.25µ ±  1%  -3.16% (p=0.000 n=15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-2          34.23µ ±   1%   33.18µ ±  1%  -3.06% (p=0.000 n=15)
MultiLineStringIsSimpleManyLineStrings/n=100-2                     25.88µ ±   3%   23.58µ ±  1%  -8.91% (p=0.000 n=15)
MultiLineStringIsSimpleManyLineStrings/n=1000-2                    322.1µ ±   1%   305.4µ ±  1%  -5.17% (p=0.000 n=15)
ForceCWandForceCCW/0-2                                             16.03n ± 539%   16.06n ±  1%       ~ (p=0.751 n=15)
ForceCWandForceCCW/0#01-2                                          102.9n ±  84%   101.2n ±  1%  -1.65% (p=0.032 n=15)
ForceCWandForceCCW/1-2                                             16.18n ± 537%   16.23n ±  1%       ~ (p=0.751 n=15)
ForceCWandForceCCW/1#01-2                                          102.7n ±  84%   101.0n ±  2%       ~ (p=0.139 n=15)
ForceCWandForceCCW/2-2                                             29.63n ±   0%   29.69n ±  1%       ~ (p=0.767 n=15)
ForceCWandForceCCW/2#01-2                                          158.2n ±   1%   155.6n ±  1%  -1.64% (p=0.024 n=15)
ForceCWandForceCCW/3-2                                             29.72n ±   0%   29.74n ±  0%       ~ (p=0.675 n=15)
ForceCWandForceCCW/3#01-2                                          158.7n ±   1%   155.5n ±  1%  -2.02% (p=0.003 n=15)
ForceCWandForceCCW/4-2                                             43.19n ±   0%   43.19n ±  0%       ~ (p=0.862 n=15)
ForceCWandForceCCW/4#01-2                                          244.3n ±   0%   240.6n ±  1%  -1.51% (p=0.017 n=15)
ForceCWandForceCCW/5-2                                             43.28n ±   0%   43.25n ±  1%       ~ (p=0.660 n=15)
ForceCWandForceCCW/5#01-2                                          246.4n ±   1%   240.0n ±  1%  -2.60% (p=0.000 n=15)
ForceCWandForceCCW/6-2                                             62.44n ±   1%   62.36n ±  0%       ~ (p=0.560 n=15)
ForceCWandForceCCW/6#01-2                                          393.0n ±   1%   386.1n ±  1%  -1.76% (p=0.002 n=15)
EnvelopeTransformXY-2                                              139.4n ±   1%   140.1n ±  0%  +0.50% (p=0.037 n=15)
WKBParse/0-2                                                       421.2n ±   1%   420.0n ±  1%       ~ (p=0.493 n=15)
WKBParse/1-2                                                       417.0n ±   1%   417.6n ±  0%       ~ (p=0.645 n=15)
WKBParse/2-2                                                       482.1n ±   1%   482.0n ±  1%       ~ (p=0.442 n=15)
WKBParse/3-2                                                       484.0n ±   1%   483.4n ±  0%       ~ (p=0.198 n=15)
WKBParse/4-2                                                       568.3n ±   1%   557.5n ±  1%  -1.90% (p=0.000 n=15)
WKBParse/5-2                                                       424.1n ±   1%   424.8n ±  0%       ~ (p=0.927 n=15)
WKBParse/6-2                                                       496.1n ±   1%   496.9n ±  1%       ~ (p=0.736 n=15)
WKBParse/7-2                                                       485.7n ±   1%   484.9n ±  0%       ~ (p=0.183 n=15)
WKBParse/8-2                                                       565.9n ±   1%   579.8n ±  1%  +2.46% (p=0.000 n=15)
WKBParse/9-2                                                       310.7n ±   2%   310.0n ±  1%       ~ (p=0.319 n=15)
WKBParse/10-2                                                      312.2n ±   1%   312.2n ±  1%       ~ (p=0.660 n=15)
WKBParse/11-2                                                      312.4n ±   1%   310.9n ±  0%  -0.48% (p=0.046 n=15)
WKBParse/12-2                                                      307.0n ±   1%   306.3n ±  0%  -0.23% (p=0.030 n=15)
WKBParse/13-2                                                      594.5n ±   1%   594.0n ±  0%       ~ (p=0.299 n=15)
WKBParse/14-2                                                      703.9n ±   2%   705.0n ±  1%       ~ (p=0.616 n=15)
WKBParse/15-2                                                      699.2n ±   2%   711.5n ±  1%  +1.76% (p=0.003 n=15)
WKBParse/16-2                                                      847.5n ±   1%   846.6n ±  0%       ~ (p=0.560 n=15)
WKBParse/17-2                                                      700.2n ±   1%   700.7n ±  1%       ~ (p=0.976 n=15)
WKBParse/18-2                                                      897.1n ±   1%   918.8n ±  2%  +2.42% (p=0.000 n=15)
WKBParse/19-2                                                      909.1n ±   1%   906.9n ±  1%       ~ (p=0.894 n=15)
WKBParse/20-2                                                      1.101µ ±   0%   1.091µ ±  0%  -0.91% (p=0.003 n=15)
WKBParse/21-2                                                      302.5n ±   1%   300.8n ±  1%       ~ (p=0.056 n=15)
WKBParse/22-2                                                      303.3n ±   1%   302.8n ±  1%       ~ (p=0.134 n=15)
WKBParse/23-2                                                      302.6n ±   1%   302.3n ±  0%       ~ (p=0.736 n=15)
WKBParse/24-2                                                      298.3n ±   1%   297.8n ±  0%       ~ (p=0.213 n=15)
WKBParse/25-2                                                      1.562µ ±   1%   1.569µ ±  1%       ~ (p=0.220 n=15)
WKBParse/26-2                                                      1.978µ ±   1%   1.979µ ±  1%       ~ (p=0.675 n=15)
WKBParse/27-2                                                      1.980µ ±   0%   1.979µ ±  0%       ~ (p=0.878 n=15)
WKBParse/28-2                                                      2.468µ ±   0%   2.457µ ±  0%  -0.45% (p=0.000 n=15)
WKBParse/29-2                                                      302.2n ±   1%   300.6n ±  1%       ~ (p=0.220 n=15)
WKBParse/30-2                                                      303.0n ±   1%   303.5n ±  1%       ~ (p=0.878 n=15)
WKBParse/31-2                                                      302.0n ±   1%   302.9n ±  1%       ~ (p=0.967 n=15)
WKBParse/32-2                                                      298.9n ±   1%   297.7n ±  1%       ~ (p=0.220 n=15)
WKBParse/33-2                                                      571.0n ±   0%   569.3n ±  1%       ~ (p=0.140 n=15)
WKBParse/34-2                                                      651.3n ±   1%   664.6n ±  1%  +2.04% (p=0.000 n=15)
WKBParse/35-2                                                      651.6n ±   1%   650.7n ±  1%       ~ (p=0.735 n=15)
WKBParse/36-2                                                      722.9n ±   0%   713.1n ±  1%  -1.36% (p=0.000 n=15)
WKBParse/37-2                                                      821.3n ±   1%   809.3n ±  1%  -1.46% (p=0.001 n=15)
WKBParse/38-2                                                      969.3n ±   0%   968.7n ±  1%       ~ (p=0.532 n=15)
WKBParse/39-2                                                      969.0n ±   1%   965.2n ±  2%       ~ (p=0.177 n=15)
WKBParse/40-2                                                      1.081µ ±   1%   1.080µ ±  0%       ~ (p=0.559 n=15)
WKBParse/41-2                                                      301.7n ±   1%   301.0n ±  0%       ~ (p=0.087 n=15)
WKBParse/42-2                                                      303.2n ±   1%   303.7n ±  1%       ~ (p=0.690 n=15)
WKBParse/43-2                                                      303.9n ±   0%   303.2n ±  0%       ~ (p=0.198 n=15)
WKBParse/44-2                                                      298.8n ±   1%   297.8n ±  0%       ~ (p=0.145 n=15)
WKBParse/45-2                                                      889.0n ±   1%   883.8n ±  1%       ~ (p=0.164 n=15)
WKBParse/46-2                                                      1.080µ ±   2%   1.075µ ±  1%       ~ (p=0.079 n=15)
WKBParse/47-2                                                      1.082µ ±   0%   1.084µ ±  1%       ~ (p=0.545 n=15)
WKBParse/48-2                                                      1.241µ ±   1%   1.235µ ±  0%       ~ (p=0.339 n=15)
WKBParse/49-2                                                      1.268µ ±   0%   1.266µ ±  1%       ~ (p=0.894 n=15)
WKBParse/50-2                                                      1.607µ ±   1%   1.607µ ±  1%       ~ (p=0.959 n=15)
WKBParse/51-2                                                      1.599µ ±   1%   1.599µ ±  0%       ~ (p=0.798 n=15)
WKBParse/52-2                                                      1.871µ ±   1%   1.866µ ±  1%       ~ (p=0.417 n=15)
WKBParse/53-2                                                      301.1n ±   2%   302.3n ±  1%       ~ (p=0.299 n=15)
WKBParse/54-2                                                      303.2n ±   1%   303.7n ±  1%       ~ (p=0.467 n=15)
WKBParse/55-2                                                      303.6n ±   2%   303.8n ±  1%       ~ (p=0.645 n=15)
WKBParse/56-2                                                      299.3n ±   1%   298.7n ±  1%       ~ (p=0.319 n=15)
WKBParse/57-2                                                      1.926µ ±   0%   1.925µ ±  0%       ~ (p=0.573 n=15)
WKBParse/58-2                                                      2.362µ ±   1%   2.370µ ±  0%       ~ (p=0.254 n=15)
WKBParse/59-2                                                      2.367µ ±   0%   2.366µ ±  0%       ~ (p=0.453 n=15)
WKBParse/60-2                                                      2.867µ ±   1%   2.860µ ±  1%       ~ (p=0.319 n=15)
WKBParse/61-2                                                      300.7n ±   1%   300.6n ±  1%       ~ (p=0.493 n=15)
WKBParse/62-2                                                      302.2n ±   1%   302.0n ±  1%       ~ (p=0.830 n=15)
WKBParse/63-2                                                      302.1n ±   2%   301.7n ±  1%       ~ (p=0.361 n=15)
WKBParse/64-2                                                      298.5n ±   1%   296.6n ±  1%  -0.64% (p=0.009 n=15)
WKBParse/65-2                                                      877.7n ±   1%   874.7n ±  2%       ~ (p=0.661 n=15)
WKBParse/66-2                                                      1.016µ ±   1%   1.028µ ±  1%  +1.18% (p=0.000 n=15)
WKBParse/67-2                                                      1.029µ ±   1%   1.015µ ±  1%  -1.36% (p=0.017 n=15)
WKBParse/68-2                                                      1.141µ ±   0%   1.139µ ±  1%       ~ (p=1.000 n=15)
geomean                                                            1.923µ          1.907µ        -0.83%

                                                           │ /tmp/tmp.9ynJa1Wrvi │          /tmp/tmp.fiLFr6Ksvk           │
                                                           │        B/op         │     B/op      vs base                  │
LineEnvelope/0-2                                                    0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=15) ¹
LineEnvelope/1-2                                                    0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=15) ¹
LineEnvelope/2-2                                                    0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=15) ¹
LineEnvelope/3-2                                                    0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=15) ¹
MarshalWKB/polygon/n=10-2                                           232.0 ± 0%       232.0 ± 0%        ~ (p=1.000 n=15) ¹
MarshalWKB/polygon/n=100-2                                        1.789Ki ± 0%     1.789Ki ± 0%        ~ (p=1.000 n=15) ¹
MarshalWKB/polygon/n=1000-2                                       16.04Ki ± 0%     16.04Ki ± 0%        ~ (p=1.000 n=15) ¹
…